[Front-Modify] comment, viewer 컴포넌트 분리
