+++ simplesns-frontend/src/components/Post/components/TuiViewer.tsx
@@ -1,7 +1,7 @@
 import "@toast-ui/editor/dist/toastui-editor.css"; // Editor's Style
 import axios from "axios";
 import "codemirror/lib/codemirror.css";
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { useDispatch } from "react-redux";
 import { useHistory } from "react-router-dom";
 import styled from "styled-components";
@@ -9,6 +9,7 @@ import { BASE_URL } from "../../../config/config.json";
 import { CommentsData } from "../../../model/ArticleModel";
 import { actions } from "../../../redux/reducer/commentReducer";
 import { config } from "../../../utils/util";
+import TuiViewerComment from "./TuiViewerComment";
 
 interface ITuiViewer {
   id: number;
@@ -30,12 +31,6 @@ export default function TuiViewer({
   const history = useHistory();
   const dispatch = useDispatch();
   const [inputComment, setInputComment] = useState<string>("");
-  const [changedComment, setChangedComment] = useState<string>("");
-  const [changeToggle, setChangeToggle] = useState<boolean>(false);
-  const [toggleId, setToggleId] = useState<number>(0);
-  const [tokenCheck, setTokenCheck] = useState<boolean>(false);
-  const [tokenNick, setTokenNick] = useState<string>("");
-  const [tokenId, setTokenId] = useState<number>(0);
 
   const deletePost = async () => {
     const _postId = id;
@@ -54,37 +49,9 @@ export default function TuiViewer({
         history.push("/");
         return window.location.reload();
       }
-
-      console.log("delete REsponse!!!", response);
     } catch (error) {
       console.error(error);
       alert("Error:!");
-    }
-
-    console.log("id", _postId);
-  };
-
-  useEffect(() => {
-    checkTokenForBtn();
-  }, []);
-
-  const checkTokenForBtn = async () => {
-    try {
-      const response = await axios.get(`${BASE_URL}/auth/token`, config);
-      if (
-        response.status === 200 &&
-        sessionStorage.getItem("token") &&
-        sessionStorage.getItem("nick") === response.data._nick
-      ) {
-        setTokenNick(response.data._nick);
-        setTokenId(response.data._id);
-        setTokenCheck(true);
-      } else {
-        setTokenCheck(false);
-      }
-    } catch (error) {
-      setTokenCheck(false);
-      console.error(error);
     }
   };
 
@@ -126,91 +93,13 @@ export default function TuiViewer({
             <i className="far fa-comment fa-2x"></i>
           </CommentButton>
         </InputWrapper>
-        {comments.map(({ comment, User, id }) => (
-          <>
-            {User.id === tokenId && changeToggle && id === toggleId ? (
-              <ChangeCommentContainer>
-                <CommentsWrapper>
-                  <span>{User.nick}</span>
-                  <input
-                    autoFocus
-                    value={changedComment}
-                    onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-                      setChangedComment(e.target.value);
-                    }}
-                  />
-                </CommentsWrapper>
-
-                <CommentsButtonWrapper>
-                  <button
-                    onClick={() => {
-                      dispatch(actions.putCommentData(id, changedComment));
-                    }}
-                  >
-                    <i className="fas fa-file-invoice"></i>
-                  </button>
-                  <button
-                    onClick={() => {
-                      setChangeToggle(false);
-                    }}
-                  >
-                    <i className="fas fa-times-circle"></i>
-                  </button>
-                </CommentsButtonWrapper>
-              </ChangeCommentContainer>
-            ) : (
-              <CommentsContainer key={User.id + "0" + id}>
-                <CommentsWrapper>
-                  <span>{User.nick}</span>
-                  <span>{comment}</span>
-                </CommentsWrapper>
-                <CommentsButtonWrapper>
-                  {User.nick === tokenNick &&
-                    User.id === tokenId &&
-                    tokenCheck && (
-                      <>
-                        <button
-                          onClick={() => {
-                            setToggleId(id);
-                            setChangeToggle(true);
-                            setChangedComment(comment);
-                          }}
-                        >
-                          <i className="fas fa-edit "></i>
-                        </button>
-                        <button
-                          onClick={() => {
-                            dispatch(actions.deleteCommentData(id));
-                          }}
-                        >
-                          <i className="fas fa-times "></i>
-                        </button>
-                      </>
-                    )}
-                </CommentsButtonWrapper>
-              </CommentsContainer>
-            )}
-          </>
-        ))}
+        <>
+          <TuiViewerComment comments={comments} />
+        </>
       </ViewerBottom>
     </ViewerContainer>
   );
 }
-
-const ChangeCommentContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0vw 0.5vw;
-
-  input {
-    border: 1px solid white;
-    border-radius: 8px;
-    margin-left: 1vw;
-    font-size: 1rem;
-    font-family: "Newsreader", serif;
-  }
-`;
 
 const ViewerContainer = styled.article`
   width: 33.333%;
@@ -270,50 +159,6 @@ const InputWrapper = styled.div`
     height: 4vh;
     width: 90%;
     border-radius: 8px;
-  }
-`;
-
-const CommentsContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0vw 0.5vw;
-`;
-
-const CommentsWrapper = styled.div`
-  /* width: 85%; */
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-  /* margin: 0 0.5vw; */
-  span {
-    line-height: 2;
-    &:first-child {
-      font-size: 1rem;
-      font-family: "Newsreader", serif;
-      font-weight: bolder;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-      overflow: hidden;
-    }
-
-    &:last-child {
-      margin-left: 1vw;
-      font-size: 1rem;
-      font-family: "Newsreader", serif;
-    }
-  }
-`;
-
-const CommentsButtonWrapper = styled.div`
-  button {
-    background-color: white;
-    border: none;
-
-    &:hover {
-      color: gray;
-      transition: all 0.3s ease-in-out;
-    }
   }
 `;
